Move OlfactoryError to the shell package
It's not used or returned by anything else in olfactor
